Modifying the firmware for testing purposes.
Done this for testing how Linux and Windows behave when receiving a valid value for one axis but invalid for another.

In Linux, the pointer moved to the right-most position of the screen (which is consistent with the behavior I previously described), not ignoring the out-of-range value.

In Windows, the pointer moved to the left-most position of the screen. It means Windows can't send a movement of only one axis, and fills the other axis with zero when receiving an invalid value. This behavior doesn't seem very helpful.